Added debug brakes

+++ Assets/Scripts/Traffic/VehicleComponents/DriveVehicle/VehicleBrakesComponent.cs
@@ -6,5 +6,5 @@ using Unity.Mathematics;
 [GenerateAuthoringComponent]
 public struct VehicleBrakesComponent : IComponentData
 {
-    [UnityEngine.HideInInspector] public float brakesUsage;
+    [UnityEngine.HideInInspector] public int brakesUsage;
 }
